Adapt some specs

+++ app/models/order_item.rb
@@ -2,7 +2,6 @@ class OrderItem < ApplicationRecord
   belongs_to :order
   belongs_to :product
 
-
   validates :order, presence: true
   validates :product, presence: true
   validates :size, presence: true
@@ -10,10 +9,12 @@ class OrderItem < ApplicationRecord
   def remove_from_cart!
     ActiveRecord::Base.transaction do
       order.reload
+      return false unless order.open?
       order.sub_total -= (self.product.price * self.quantity)
       order.save!
       self.destroy!
     end
+    true
   end
 
 end
